refactor: Table factories
