+++ _P1/Logic/src/main/java/vdm/p1/logic/objects/LifeManager.java
@@ -1,45 +1,41 @@
 package vdm.p1.logic.objects;
 
 import vdm.p1.engine.IEngine;
-import vdm.p1.engine.IImage;
 import vdm.p1.logic.GameObject;
 import vdm.p1.logic.layout.HorizontalAlignment;
 import vdm.p1.logic.layout.Padding;
 import vdm.p1.logic.layout.VerticalAlignment;
 
 public class LifeManager extends GameObject {
+	private final IEngine engine;
+	private int remainingLives;
 
-	IEngine engine;
-	int life_number;
-	public LifeManager(IEngine engine){
-		this.engine=engine;
-		this.life_number=3;
-		GameObject heart_left = new Image(engine.getGraphics().newImage("image/heartfill.png"))
+	public LifeManager(IEngine engine) {
+		this.engine = engine;
+		this.remainingLives = 3;
+
+		GameObject heartLeft = new Image(engine.getGraphics().newImage("image/heartfill.png"))
 				.setHorizontalAlignment(HorizontalAlignment.LEFT)
 				.setVerticalAlignment(VerticalAlignment.TOP);
 
-
-		GameObject heart_center = new Image(engine.getGraphics().newImage("image/heartfill.png"))
+		GameObject heartCenter = new Image(engine.getGraphics().newImage("image/heartfill.png"))
 				.setHorizontalAlignment(HorizontalAlignment.CENTRE)
 				.setVerticalAlignment(VerticalAlignment.TOP);
 
-
-		GameObject heart_right = new Image(engine.getGraphics().newImage("image/heartfill.png"))
+		GameObject heartRight = new Image(engine.getGraphics().newImage("image/heartfill.png"))
 				.setHorizontalAlignment(HorizontalAlignment.RIGHT)
 				.setVerticalAlignment(VerticalAlignment.TOP);
 
-
-		GameObject padding= new Padding(0.04,0.1)
-				.addChild(heart_left)
-				.addChild(heart_center)
-				.addChild(heart_right);
+		GameObject padding = new Padding(0.04, 0.1)
+				.addChild(heartLeft)
+				.addChild(heartCenter)
+				.addChild(heartRight);
 
 		addChild(padding);
 	}
 
-	public boolean removeHeart(){
-		life_number--;
-		GameObject pass
-		return life_number > 0;
+	public boolean removeHeart() {
+		//GameObject pass
+		return --remainingLives > 0;
 	}
 }
