+++ _P1/Logic/src/main/java/vdm/p1/logic/scenes/GameScene.java
@@ -12,24 +12,22 @@ import vdm.p1.logic.layout.VerticalAlignment;
 import vdm.p1.logic.objects.CheckSolutionButton;
 import vdm.p1.logic.objects.GoToStartSceneButton;
 import vdm.p1.logic.objects.Image;
-import vdm.p1.logic.objects.RandomTable;
-import vdm.p1.logic.objects.ReadingTable;
+import vdm.p1.logic.objects.Table;
 import vdm.p1.logic.objects.Text;
 
 public final class GameScene extends Scene {
 	private final ISound sound;
 
-	public GameScene(IEngine engine, String filename){
+	public GameScene(IEngine engine, String filename) {
 		super(engine);
 		IFont font = engine.getGraphics().newFont("font/pico.ttf", 20, true);
 		sound = engine.getAudio().createSound("audio/meadow_thoughts");
 
-		ReadingTable table= (ReadingTable) new ReadingTable(font,filename)
+		Table table = (Table) Table.fromFile(font, filename)
 				.setHorizontalAlignment(HorizontalAlignment.CENTRE)
 				.setVerticalAlignment(VerticalAlignment.BOTTOM);
 
 		GameObject header = new Padding(0, 0, 0.8, 0);
-
 
 		GameObject padding = new Padding(0.04, 0.1)
 				.addChild(header)
@@ -49,7 +47,7 @@ public final class GameScene extends Scene {
 		super(engine);
 
 		IFont font = engine.getGraphics().newFont("font/pico.ttf", 20, true);
-		RandomTable table = (RandomTable) new RandomTable(font, rows, columns)
+		Table table = (Table) Table.fromRandom(font, rows, columns)
 				.setHorizontalAlignment(HorizontalAlignment.CENTRE)
 				.setVerticalAlignment(VerticalAlignment.BOTTOM);
