+++ _P1/app/src/main/Assets/levels/5x5_01
@@ -1,0 +1,6 @@
+5 5 Pez
+_ _ O _ _
+O _ O O _
+O O O O O
+O _ O O _
+_ _ O _ _
